refactor: merge disposable implementation into base instance
