+++ src/index.ts
@@ -51,21 +51,14 @@ export type Options = {
   docInitParams?: Partial<DocumentInitParameters>;
 };
 
-interface BasePdf {
+export interface Pdf extends AsyncDisposable {
   length: number;
   metadata: PdfMetadata;
   isDestroyed: boolean;
   getPage(pageNumber: number): Promise<Buffer>;
   destroy(): Promise<void>;
-  [Symbol.asyncIterator](): AsyncIterator<Buffer, void, void>;
-}
-
-export interface Pdf extends BasePdf {
-  asAsyncDisposable(): DisposablePdf;
-}
-
-export interface DisposablePdf extends BasePdf, AsyncDisposable {
   [Symbol.asyncDispose](): Promise<void>;
+  [Symbol.asyncIterator](): AsyncIterator<Buffer, void, void>;
 }
 
 /**
@@ -140,13 +133,15 @@ export async function pdf(
     isDestroyed = true;
   };
 
-  // `isDestroyed` is added by each caller because re-using the `get` results
-  // in freezing the value at the time of spread, since spread invokes accessors.
-  const base: Omit<BasePdf, "isDestroyed"> = {
+  return {
     length: pdfDocument.numPages,
     metadata: sanitize(metadata.info),
     getPage,
+    get isDestroyed(): boolean {
+      return isDestroyed;
+    },
     destroy,
+    [Symbol.asyncDispose]: destroy,
     [Symbol.asyncIterator]() {
       return {
         pg: 0,
@@ -160,19 +155,5 @@ export async function pdf(
         },
       };
     },
-  };
-
-  return {
-    ...base,
-    get isDestroyed(): boolean {
-      return isDestroyed;
-    },
-    asAsyncDisposable: (): DisposablePdf => ({
-      ...base,
-      get isDestroyed(): boolean {
-        return isDestroyed;
-      },
-      [Symbol.asyncDispose]: destroy,
-    }),
-  };
+  } satisfies Pdf;
 }
